chore: move projen tests to src/__tests__
Move tests to `src/__tests__` so they are executed after compilation. This enables quicker turn around time by using `watch` in the background.
