+++ .projenrc.js
@@ -13,6 +13,8 @@ const project = new JsiiProject({
     'By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.',
   ],
 
+  testdir: 'src/__tests__',
+
   bundledDeps: [
     'yaml',
     'fs-extra',
