+++ package.json
@@ -58,7 +58,6 @@
     "json-schema": "^0.2.5",
     "markmac": "^0.1.23",
     "standard-version": "^9.0.0",
-    "ts-jest": "^26.4.4",
     "typescript": "^3.9.5"
   },
   "dependencies": {
@@ -91,8 +90,7 @@
   "version": "0.13.1",
   "jest": {
     "testMatch": [
-      "**/__tests__/**/*.ts?(x)",
-      "**/?(*.)+(spec|test).ts?(x)"
+      "**/lib/__tests__/**/?(*.)+(spec|test).js?(x)"
     ],
     "clearMocks": true,
     "collectCoverage": true,
@@ -111,13 +109,7 @@
           "outputDirectory": "test-reports"
         }
       ]
-    ],
-    "preset": "ts-jest",
-    "globals": {
-      "ts-jest": {
-        "tsconfig": "tsconfig.jest.json"
-      }
-    }
+    ]
   },
   "types": "lib/index.d.ts",
   "stability": "experimental",
