+++ src/__tests__/cli.test.ts
