+++ src/__tests__/docker-compose.test.ts
@@ -1,8 +1,8 @@
 import * as child_process from 'child_process';
 import * as path from 'path';
 import * as fs from 'fs-extra';
-import { DockerCompose, DockerComposeProtocol } from '../src';
-import * as logging from '../src/logging';
+import { DockerCompose, DockerComposeProtocol } from '..';
+import * as logging from '../logging';
 import { TestProject } from './util';
 
 logging.disable();
