+++ src/__tests__/eslint.test.ts
@@ -1,9 +1,14 @@
-import { Eslint, NodeProject } from '../src';
+import { Eslint, NodeProject } from '..';
+import { LogLevel } from '../logger';
 import { mkdtemp, synthSnapshot } from './util';
 
 test('devdirs', () => {
   // GIVEN
-  const project = new NodeProject({ outdir: mkdtemp(), name: 'test' });
+  const project = new NodeProject({
+    outdir: mkdtemp(),
+    name: 'test',
+    logging: { level: LogLevel.OFF },
+  });
 
   // WHEN
   new Eslint(project, {
