+++ src/__tests__/ignore-file.test.ts
@@ -1,4 +1,4 @@
-import { IgnoreFile } from '../src';
+import { IgnoreFile } from '..';
 import { synthSnapshot, TestProject } from './util';
 
 test('ignorefile synthesises correctly', () => {
