+++ src/__tests__/new.test.ts
@@ -1,13 +1,11 @@
 // tests for `projen new`: we run `projen new` for each supported project type
 // and compare against a golden snapshot.
-import { execSync, spawnSync } from 'child_process';
+import { execSync } from 'child_process';
 import { join } from 'path';
 import { mkdirSync, readFileSync, removeSync } from 'fs-extra';
-import { PROJEN_RC } from '../src/common';
-import * as inventory from '../src/inventory';
-import { mkdtemp, synthSnapshot, synthSnapshotWithPost, TestProject } from './util';
-
-const cli = require.resolve('../bin/projen');
+import { PROJEN_RC } from '../common';
+import * as inventory from '../inventory';
+import { execProjenCLI, mkdtemp, synthSnapshot, synthSnapshotWithPost, TestProject } from './util';
 
 for (const type of inventory.discover()) {
   test(`projen new ${type.pjid}`, () => {
@@ -16,11 +14,7 @@ for (const type of inventory.discover()) {
       const projectdir = createProjectDir(outdir);
 
       // execute `projen new PJID --no-synth` in the project directory
-      const ret = spawnSync(process.execPath, [cli, 'new', '--no-synth', type.pjid], { cwd: projectdir });
-
-      // throw on errors
-      if (ret.error) { throw new Error(ret.error.message); }
-      if (ret.status !== 0) { throw new Error(ret.stderr.toString('utf-8')); }
+      execProjenCLI(projectdir, ['new', '--no-synth', type.pjid]);
 
       // compare generated .projenrc.js to the snapshot
       const projenrc = readFileSync(join(projectdir, PROJEN_RC), 'utf-8');
