+++ src/__tests__/vscode-launch-config.test.ts
@@ -1,4 +1,4 @@
-import { InternalConsoleOptions } from '../src/vscode/launch-config';
+import { InternalConsoleOptions } from '../vscode/launch-config';
 import { synthSnapshot, TestProject } from './util';
 
 const VSCODE_DEBUGGER_FILE = '.vscode/launch.json';
